Added new cli argument --kw for status command
It looks like the Renault API is inconsistent in reporting chargingInstantaneousPower.
The CLI assumes that it is reported as Watt, but in my case it is reported as kiloWatt.
The result is that the value for Charge Rate is to small by the factor 1000.

Therefore I added a new cli argument --kw for the status command which interprets chargingInstantaneousPower as kW.
The default behaviour without --km argument the same as the current behaviour.